requirements.txt
discord.py==1.4.1
peewee==3.13.3
psycopg2-binary==2.8.5
